requirements.txt
# Core Dependencies - Direct imports only
# Bot Framework
nonebot2>=2.4.4
nonebot-adapter-onebot>=2.4.6
nonebot-plugin-apscheduler~=0.5.0

# HTTP Clients & Web
aiohttp>=3.10.3,<3.11.0
httpx>=0.28.1
requests>=2.32.5
urllib3>=2.6.3

# Data Processing & Analysis
numpy>=2.2.6
scipy>=1.14.1
matplotlib>=3.10.8
wordcloud>=1.9.6
jieba~=0.42.1

# Image Processing
pillow>=12.1.0

# Parsing & Serialization
lxml>=6.0.2
markdown2>=2.5.4
pure-protobuf>=3.1.5

# API Clients
openai>=2.17.0
pixivpy3>=3.7.5

# Browser Automation
playwright>=1.58.0

# Utilities
loguru>=0.7.3
python-dotenv>=1.2.1
pydantic~=2.8.2
typing_extensions~=4.12.2
yarl>=1.22.0
tqdm>=4.67.3
aiocache>=0.12.3

# Media Download
youtube-dl~=2021.12.17

# Compression
Brotli>=1.1.0

# NoneBot2 Core Dependencies (explicitly listed for clarity)
pygtrie~=2.5.0
tomli>=2.4.0

# APScheduler (required by nonebot-plugin-apscheduler)
APScheduler>=3.11.2

